Je chie sur les ftp, jupyter et tout ces trucs deprimants

+++ 5-onycroit/srcs/requirements/bonus/jupyter/tools/jupyterstart.sh
@@ -1,9 +1,8 @@
 echo "Launching jupyter"
 
 jupyter notebook --allow-root --generate-config
-#[ -f /tmp/jupyter_notebook_config.py ] && \
-mv /tmp/jupyter_notebook_config.py /root/.jupyter/jupyter_notebook_config.py
+[ -f /tmp/jupyter_notebook_config.py ] && mv /tmp/jupyter_notebook_config.py /root/.jupyter/jupyter_notebook_config.py
 
-jupyter notebook --no-browser --allow-root --ip 0.0.0.0
+#jupyter notebook --no-browser --allow-root --ip 0.0.0.0
 
-#jupyter notebook
+jupyter notebook
